# bar.openapi.yml
openapi: 3.0.3
info:
  title: OpenAPI - Bar
  license:
    name: MIT
    url: https://opensource.org/license/mit/
  version: 1.0.11
servers:
  - url: http://localhost:8080
  - url: https://le-bar.telecomnancy.net/api/v1
...
paths:
...
  /categories:
    get:
      description: Get all categories
      operationId: getCategories
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - auth: []
      tags:
        - categories
    post:
      description: Create a new category
      operationId: postCategory
      requestBody:
        description: Category object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCategory'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '401':
          description: 'Not authorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '409':
          description: 'Category already exists'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - admin_auth: []
      tags:
        - categories
  /categories/{category_id}:
    get:
      description: Get a category
      operationId: getCategory
      parameters:
        - name: category_id
          in: path
          description: ID of the category
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          description: 'Category not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - auth: []
      tags:
        - categories
    patch:
      description: Update a category
      operationId: patchCategory
      parameters:
        - name: category_id
          in: path
          description: ID of the category
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        description: Category object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategory'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '401':
          description: 'Not authorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: 'Category not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - admin_auth: []
      tags:
        - categories
    delete:
      description: Delete a category
      operationId: markDeleteCategory
      parameters:
        - name: category_id
          in: path
          description: ID of the category
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '204':
          description: ''
        '404':
          description: 'Category not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - admin_auth: []
      tags:
        - categories
...
  /categories/{category_id}/items:
    get:
      description: Get all items of a category
      operationId: getCategoryItems
      parameters:
        - name: category_id
          in: path
          description: ID of the category
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            format: uint64
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: uint64
        - name: state
          in: query
          description: Filter by state
          required: false
          schema:
            type: string
            $ref: '#/components/schemas/ItemState'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Item'
                  page:
                    type: integer
                    format: uint64
                    
                  limit:
                    type: integer
                    format: uint64
                    
                  max_page:
                    type: integer
                    format: uint64
                    
        '404':
          description: 'Category not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError' 
      security:
        - auth: []
      tags:
        - items
    post:
      description: Create a new item
      operationId: postItem
      parameters:
        - name: category_id
          in: path
          description: ID of the category
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        description: Item object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewItem'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '401':
          description: 'Not authorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '409':
          description: 'Item already exists'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - admin_auth: []
      tags:
        - items
  /categories/{category_id}/items/{item_id}:
    patch:
      description: Update an item
      operationId: patchItem
      parameters:
        - name: category_id
          in: path
          description: ID of the category
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: item_id
          in: path
          description: ID of the item
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        description: Item object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItem'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '401':
          description: 'Not authorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: 'Item not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - admin_auth: []
      tags:
        - items
    delete:
      description: Delete an item
      operationId: markDeleteItem
      parameters:
        - name: category_id
          in: path
          description: ID of the category
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: item_id
          in: path
          description: ID of the item
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '204':
          description: ''
        '401':
          description: 'Not authorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: 'Item not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - admin_auth: []
      tags:
        - items
  /categories/{category_id}/items/{item_id}/picture:
    get:
      description: 'Get an item picture'
      operationId: getItemPicture
      parameters:
        - name: category_id
          in: path
          description: ID of the category
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: item_id
          in: path
          description: ID of the item
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: ''
          content:
            image/png:
              schema:
                type: string
                format: binary
        '404':
          description: 'Item not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
        - auth: []
      tags:
        - items
...
components:
  schemas:
    UUID:
      type: string
      format: uuid
      example: 4be0643f-1d98-573b-97cd-ca98a65347dd
    ErrorCodes:
      type: string
      enum:
        - bad_request
        - forbidden
        - account_not_found
        - qr_invalid
        - internal_server_error
        - not_authenticated
        - image_not_found
        - text_not_found
        - category_not_found
        - item_not_found
        - refill_not_found
        - transaction_not_found
      x-enum-varnames:
        - ErrBadRequest
        - ErrForbidden
        - ErrAccountNotFound
        - ErrQRInvalid
        - ErrInternalServerError
        - ErrNotAuthenticated
        - ErrImageNotFound
        - ErrTextNotFound
        - ErrCategoryNotFound
        - ErrItemNotFound
        - ErrRefillNotFound
        - ErrTransactionNotFound
    Messages:
      type: string
      enum:
        - Bad request
        - Not connected
        - Missing fields %v
        - Account already exists
        - Account not found
        - Account is not allowed to use ressource
        - Account cannot use ressource at the time being
        - Invalid QR Code nonce
        - Internal server error
        - You are not authenticated
        - Image does not exists
        - Text does not exists
        - Category does not exists
        - Item does not exists
        - Refill does not exists
        - Transaction does not exists
      x-enum-varnames:
        - MsgBadRequest
        - MsgNotConnected
        - MsgMissingFields
        - MsgAccountAlreadyExists
        - MsgAccountNotFound
        - MsgAccountNotAllowed
        - MsgAccountNotAvailable
        - MsgQRInvalid
        - MsgInternalServerError
        - MsgNotAuthenticated
        - MsgImageNotFound
        - MsgTextNotFound
        - MsgCategoryNotFound
        - MsgItemNotFound
        - MsgRefillNotFound
        - MsgTransactionNotFound
    HTTPError:
      type: object
      properties:
        message:
          $ref: '#/components/schemas/Messages'
        error_code:
          $ref: '#/components/schemas/ErrorCodes'
      required:
        - message
        - error_code
...
    ItemState:
      type: string
      description: State of the item
      example: buyable
      enum:
        - buyable
        - not_buyable
      x-enum-varnames:
        - ItemBuyable
        - ItemNotBuyable
    UpdateItem:
      type: object
      properties:
        price:
          type: integer
          format: uint64
          example: 1000
        amount_left:
          type: integer
          format: uint64
          example: 32
        buy_limit:
          type: integer
          format: uint64
          example: 32
        category_id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          description: Name of the current item
        picture:
          type: string
          format: base64
          description: Picture of the current item
        promotion:
          type: integer
          format: uint64
        promotion_ends_at:
          type: integer
          format: uint64
        state:
          $ref: '#/components/schemas/ItemState'
    NewItem:
      type: object
      properties:
        price:
          type: integer
          format: uint64
          example: 1000
        promotion:
          type: integer
          format: uint64
        promotion_ends_at:
          type: integer
          format: uint64
        amount_left:
          type: integer
          format: uint64
          example: 32
        buy_limit:
          type: integer
          format: uint64
          example: 32
        name:
          type: string
          description: Name of the current item
        picture:
          type: string
          format: base64
          description: Picture of the current item
        state:
          $ref: '#/components/schemas/ItemState'
      required:
        - price
        - amount_left
        - buy_limit
        - category_id
        - name
        - picture
        - state
    Item:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        price:
          type: integer
          format: uint64
          example: 1000
        promotion:
          type: integer
          format: uint64
          example: 3000 (-30%)
        promotion_ends_at:
          type: integer
          format: uint64
        amount_left:
          type: integer
          format: uint64
          example: 32
        buy_limit:
          type: integer
          format: uint64
          example: 32
        category_id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          description: Name of the current item
        picture_uri:
          type: string
          description: Link to picture of the current item
        state:
          $ref: '#/components/schemas/ItemState'
        deleted_at:
          type: integer
          format: uint64
        deleted_by:
          $ref: '#/components/schemas/UUID'
      required:
        - id
        - price
        - amount_left
        - buy_limit
        - category_id
        - name
        - picture_uri
        - state
    UpdateCategory:
      type: object
      properties:
        name:
          type: string
          description: Name of the current category
        picture:
          type: string
          format: base64
    NewCategory:
      type: object
      properties:
        name:
          type: string
          description: Name of the current category
        picture:
          type: string
          format: base64
      required:
        - name
        - picture
    Category:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          description: Name of the current category
        picture_uri:
          type: string
          description: Link to picture of the current category
        deleted_at:
          type: integer
          format: uint64
        deleted_by:
          $ref: '#/components/schemas/UUID'
      required:
        - id
        - name
        - picture_uri
        - items
...
  securitySchemes:
    auth:
      type: apiKey
      name: BAR_SESS
      in: cookie
    local_token:
      type: apiKey
      name: X-Local-Token
      in: header
    admin_auth:
      type: apiKey
      name: BAR_ADMIN_SESS
      in: cookie